+++ manta-accounting/src/wallet/mod.rs
@@ -37,8 +37,8 @@ use crate::{
         balance::{BTreeMapBalanceState, BalanceState},
         ledger::ReadResponse,
         signer::{
-            BalanceUpdate, Checkpoint, IdentityRequest, IdentityResponse, InitialSyncData,
-            InitialSyncRequest, SignError, SignRequest, SignResponse,
+            BalanceUpdate, Checkpoint, ConsolidationPrerequest, IdentityRequest, IdentityResponse,
+            InitialSyncData, InitialSyncRequest, SignError, SignRequest, SignResponse,
             SignWithTransactionDataResponse, SyncData, SyncError, SyncRequest, SyncResponse,
             SyncResult, TransactionDataRequest, TransactionDataResponse,
         },
@@ -564,6 +564,21 @@ where
             .map_err(Error::SignError)
     }
 
+    /// Consolidates the assets in `request` using the signer connection. This
+    /// method _does not_ automatically sychronize with the ledger. To do this, call the
+    /// [`sync`](Self::sync) method separately.
+    #[inline]
+    pub async fn consolidate(
+        &mut self,
+        request: ConsolidationPrerequest<C>,
+    ) -> Result<SignResponse<C>, Error<C, L, S>> {
+        self.signer
+            .consolidate(request)
+            .await
+            .map_err(Error::SignerConnectionError)?
+            .map_err(Error::SignError)
+    }
+
     /// Attempts to process [`TransferPost`]s and returns the corresponding
     /// [`TransactionData`](crate::transfer::canonical::TransactionData).
     #[inline]
@@ -624,6 +639,42 @@ where
     {
         self.sync().await?;
         let SignResponse { posts } = self.sign(transaction, metadata).await?;
+        self.ledger
+            .write(posts)
+            .await
+            .map_err(Error::LedgerConnectionError)
+    }
+
+    /// Posts a consolidation to the ledger, returning a success [`Response`] if the assets in
+    /// `request` were successfully consolidated and posted to the ledger. This method automatically
+    /// synchronizes with the ledger before posting, _but not after_. To amortize the cost of future
+    /// calls to [`post_consolidation`], the [`sync`] method can be used to synchronize with the ledger.
+    ///
+    /// # Failure Conditions
+    ///
+    /// This method returns a [`Response`] when there were no errors in producing transfer data and
+    /// sending and receiving from the ledger, but instead the ledger just did not accept the
+    /// transaction as is. This could be caused by an external update to the ledger while the signer
+    /// was building the transaction that caused the wallet and the ledger to get out of sync. In
+    /// this case, [`post_consolidation`] can safely be called again, to retry the transaction.
+    ///
+    /// This method returns an error in any other case. The internal state of the wallet is kept
+    /// consistent between calls and recoverable errors are returned for the caller to handle.
+    ///
+    /// [`Response`]: ledger::Write::Response
+    /// [`post_consolidation`]: Self::post_consolidation
+    /// [`sync`]: Self::sync
+    #[inline]
+    pub async fn post_consolidation(
+        &mut self,
+        request: ConsolidationPrerequest<C>,
+    ) -> Result<L::Response, Error<C, L, S>>
+    where
+        L: ledger::Read<SyncData<C>, Checkpoint = S::Checkpoint>
+            + ledger::Write<Vec<TransferPost<C>>>,
+    {
+        self.sync().await?;
+        let SignResponse { posts } = self.consolidate(request).await?;
         self.ledger
             .write(posts)
             .await